+++ spinta/datasets/backends/sqldump/ufuncs/ufuncs.py
@@ -1,10 +1,17 @@
 from pathlib import Path
 
 from spinta.core.ufuncs import Bind
+from spinta.core.ufuncs import Expr
 from spinta.core.ufuncs import ufunc
 from spinta.datasets.backends.sqldump.ufuncs.components import File
 from spinta.datasets.backends.sqldump.ufuncs.components import PrepareFileResource
 from spinta.exceptions import UnknownBind
+
+
+@ufunc.resolver(PrepareFileResource, Expr)
+def file(env: PrepareFileResource, expr: Expr) -> File:
+    args, kwargs = expr.resolve(env)
+    return env.call('file', *args, **kwargs)
 
 
 @ufunc.resolver(PrepareFileResource, Bind)